refactor: move get_fullname and get_package_dir functions to pqi_path_helper
